requirements.txt
django
djangorestframework
django-storages
boto3